requirements.txt
pandas
pymongo
python-dotenv